third_party/WebKit/LayoutTests/http/tests/appcache/wrong-signature-2.html

<html manifest="resources/wrong-signature-2.manifest">
<script src="/js-test-resources/js-test.js"></script>
<script>
var jsTestIsAsync = true;
description("Test that a manifest served with a wrong signature isn't treated as such.");

function unexpectedEvent(name)
{
    testFailed("Unexpected " + name + " event.");
    finishJSTest();
}

function test(e)
{
    shouldBe("window.applicationCache.status", "applicationCache.UNCACHED");

    event = e;
    shouldBeEqualToString("event.reason", "signature");
    shouldBeEqualToString("event.url", "");
    shouldBe("event.status", "0");
    shouldBeTrue("'message' in event");

    finishJSTest();
}

applicationCache.addEventListener('cached', function() { unexpectedEvent("cached"); }, false);
applicationCache.addEventListener('noupdate', function() { unexpectedEvent("noupdate"); }, false);
applicationCache.addEventListener('error', test, false);

</script>
</body>
</html>
