third_party/WebKit/LayoutTests/http/tests/appcache/wrong-signature-2-expected.txt

Test that a manifest served with a wrong signature isn't treated as such.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.applicationCache.status is applicationCache.UNCACHED
PASS event.reason is "signature"
PASS event.url is ""
PASS event.status is 0
PASS 'message' in event is true
PASS successfullyParsed is true

TEST COMPLETE
